requirements.txt
flask
gunicorn
pillow
numpy
torch
torchvision
